changes to the operator + manager classes

+++ src/main/java/org/example/Models/Employees/Operators.java
@@ -2,7 +2,7 @@ package org.example.Models.Employees;
 
 public class Operators extends Employee {
     @Override
-    public Integer getSalary() {
+    public Double getSalary() {
         return super.getSalary();
     }
